more debugging work - just getting the scripts to compile and go
